move utils for comment
